test(sec2): align money-column test schemas to the real BIGINT µ-unit type + guard (#303)
Phase-4-PRE. Recon inverted the stated finding: prod minted_amount is BIGINT (0082), not DOUBLE — no SEC-2 regression; every conserved column is BIGINT µ-units. The real divergences were 7 test files declaring conserved columns DOUBLE (stale pre-SEC-2 schemas, per-file DDL) — aligned all to BIGINT so the suite exercises the real type. Extended sec2_realpg to assert BIGINT on all 8 mint tables' minted_amount + DOUBLE on Tier-3 (divergence-catcher). Prod schema unchanged; test-only. The 0091 collision was fixed independently by #302.

+++ internal/earnverify/verify_integration_test.go
@@ -26,7 +26,7 @@ func earnTestPool(t *testing.T) *pgxpool.Pool {
 		`DROP TABLE IF EXISTS lxc_purchases`,
 		`DROP TABLE IF EXISTS workspaces`,
 		`CREATE TABLE workspaces (id TEXT PRIMARY KEY, earn_verified BOOLEAN NOT NULL DEFAULT false)`,
-		`CREATE TABLE lxc_purchases (workspace_id TEXT NOT NULL, status TEXT NOT NULL, lxc_amount DOUBLE PRECISION NOT NULL DEFAULT 0)`,
+		`CREATE TABLE lxc_purchases (workspace_id TEXT NOT NULL, status TEXT NOT NULL, lxc_amount BIGINT NOT NULL DEFAULT 0)`, // BIGINT µLXC (matches prod 0083)
 	} {
 		if _, err := pool.Exec(ctx, ddl); err != nil {
 			t.Fatalf("schema: %v", err)
